requirements.txt
flask == 3.1.0
flask_sqlalchemy == 3.1.1
flask_socketio == 5.5.1
python-whois == 0.9.5
cryptography